Add PipelineProvider
